Detect host disconnect via alt_setting reset
